+++ client/v2/CHANGELOG.md
@@ -45,7 +45,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Improvements
 
-* [#19060](https://github.com/cosmos/cosmos-sdk/pull/19060) Use clientCtx from root command in autocli commands.
+* [#19060](https://github.com/cosmos/cosmos-sdk/pull/19060) Use client context from root (or enhanced) command in autocli commands.
+  * Note, the given command must have a `client.Context` in its context.
 
 ### Bug Fixes